Move find_models_with_matching_fields to core.models.utils
